requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
streamlit>=1.29.0
plotly>=5.18.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
